GF-5[feat]: Room 생성요청 DTO 분리
